Lock mime-types to < 3 for Rails 4 & ruby 1.9
Newer versions of the mime-types gem require ruby 2.0, so tests fail on
ruby 1.9 when trying to bundle.  This fixes that for ruby 1.9, but
unfortunately seems to install mime-types 2.99 for ruby across the
board, instead of just for ruby 1.9. Not sure if this is a problem since
mime-types is something that shouldn't be affected in this project.