Update to support modern rubies and rails
This is, unfortunately, a much bigger commit than I would generally
prefer. In it, we make a _big_ jump into modern versions of rubies and
rails.

Many of the combinations in the Travis CI matrix were for EOL'd
dependencies that it no longer makes sense for this gem to support. If
you're running a Rails 3 application still... I don't know what to say.
Contact me? Let's talk about all of the security and performance
benefits you would have if you upgraded to _supported_ versions.

* Dropped ruby 2.0, 2.1

* Dropped rails 3.2, 4.0, 4.1. Should probably drop 4.2 too, but the
  tests still work so let's keep it there a bit longer. Expect this to
  be dropped too in the future, without warning, since this version of
  Rails is no longer supported.

* Added gemfiles for rails 5.1, 5.2, and 6.0! You're welcome, early
  upgraders. (But also, _how_ are you still using this gem after 3 years
  without updates??)

* Dropped development dependency on sqlite3. This was never actually
  used by the test suite for this gem, and dropping the dependency on it
  allows me to avoid dealing with various sqlite3 compatibility issues.

* Simplified all dummy applications (`test/dummy-*`) used in the test
  suite. Unused frameworks are no longer loaded, and many `--skip-*`
  flags were used to remove unused files. Removing the dependency on
  `active_record` in these dummy applications is what allowed me to remove
  the `sqlite3` dependency.

Co-authored-by: Jörg Schiller <joergschiller@googlemail.com>