load dummy for any version
This commit trades in the hard-coded Rails versions for a more dynamic
approach to loading the correct dummy application, and improves the
error message that is displayed when attempting to run the test suite on
a version of Rails that is not yet supported.

To generate a new dummy application, start with `rails plugin --help`
and review the list of available flags. Skip almost everyhing except
sprockets.